Add mockito-target-minus-junit4.
As the name suggests, this is a version of mockito-target that contains
no junit classes. Most unit tests are run with android.test.runner in their
classpath, and that library pulls in its own copy of junit that might be a
different version from the one being pulled in here.

This is pretty brittle, but we've painted ourselves into a corner by
making junit a public API anyway.

bug: 26458139

Change-Id: I5d5f7c53364264b88cdab6292eab38c147a0c737