Add mockito-target-minus-junit4.
As the name suggests, this is a version of mockito-target that contains
no junit classes. Most unit tests are run with android.test.runner in their
classpath, and that library pulls in its own copy of junit that might be a
different version from the one being pulled in here.

This is pretty brittle, but we've painted ourselves into a corner by
making junit a public API anyway.

bug: 26458139
bug: 29411540

(cherry picked from commit 865c8027e12c5e19e99f808132357a15736eb030)

Change-Id: I84f51ea2438fc507034ff806e6c76fc432b1bbf3