Fix upstream mismatch between cglib-and-asm source and library
Upstream has the source for cglib-and-asm but doesn't actually
build against that, instead it builds against a precompiled jar,
lib/repackaged/cglib-and-asm-1.0.jar. As a result of that it
appears as though they have diverged causing this problem.

This is a long standing problem that has been fixed before in
the Android version of mockito.

Bug: 30299479
Test: Run mockito tests in both CTS and Vogar
Change-Id: I12ee1a8c282962ceaa2dff4d45cd7d5fb733845d