Add classes and methods to aid in the upgrade to 2.7.13
Unsurprisingly there are a number of incompatibilities between
the current version (1.10.19) and 2.7.13. These classes and
methods were added to fix the problems found while trying to
build against 2.7.13.

The classes in org.mockito.compat were added to act as a layer
to insulate the code against those changes in mockito. The
intent is that the existing code will be changed to use those
classes instead of the ones that they extend so it builds
against 1.10.19. Then the upgrade will modify those classes
to make them work with 2.7.13 so that the code continues to
build. After the upgrade has been completed then the code that
used those compat classes will be modified to use standard
mockito and the classes will be deleted.

The MockitoHamcrest method exists upstream, this is just a
copy of that which works in 2.7.13. This is needed to support
using hamcrest based matchers with mockito.

The changes to InvocationOnMock adds a method that is new in
2.7.13 and deprecate the method that will be removed. Existing
code will be modified to use the new method and the upgrade will
delete the old one.

Bug: 32912773
Test: make checkbuild
Change-Id: I31dc4d3fc84558afda0c00f8054ec1fd4ec54c5a