Improve Mockito 2.7.13 compatibility classes
Adds describeTo(Description) method to call toString() method
to get the description. 2.7.13 does not use that method at all
because it is removing runtime dependencies on Hamcrest and so
any implementations need to be modified to update their
toString() method.

Add default toString() implementation to match 1.10.19's default
implementation of the describeTo(Description) method.

Bug: 32912773
Test: make checkbuild
Change-Id: I787fb62e931ea89899ede9db235ed156892d4b78